+++ front/src/components/NotFound/NotFound.scss
@@ -6,6 +6,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  user-select: none;
 
   .content {
     display: flex;
